requirements.txt
discord.py
asyncio
pytz
datetime
time